Moved new-payment-method to inside tokeize-payment-method
